move: BootcampService -> postSingleDiploma, postMultipleDiplomas to DiplomaService
