+++ frontend/src/services/bootcampService.tsx
@@ -1,4 +1,4 @@
-import { BootcampInDiploma, BootcampRequest, BootcampResponse, DiplomaRequest, DiplomaResponse, DiplomasRequestDto } from "../util/types";
+import { BootcampRequest, BootcampResponse } from "../util/types";
 
 export async function getBootcamps(): Promise<BootcampResponse[]>{
     const response = await fetch(`http://localhost:5258/api/bootcamp`);
@@ -63,41 +63,4 @@ export async function deleteBootcampById(guidId: string): Promise<void> {
             throw new Error('Bootcamp not found');
         throw new Error('Failed to delete bootcamp!');
     }
-}
-
-export async function postSingleDiploma(diplomaRequest: DiplomaRequest): Promise<DiplomaResponse> {
-    const response = await fetch(`http://localhost:5258/api/diploma/single`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(diplomaRequest)
-    });
-
-    if (!response.ok) {
-        if (response.status === 409) {
-            throw new Error("This student has already earned a diploma in this bootcamp!");
-        }
-        throw new Error("Failed to post new diploma!");
-    }
-
-    const result = await response.json() as DiplomaResponse;
-    return result;
-}
-
-
-export async function postMultipleDiplomas(diplomasRequest: DiplomasRequestDto): Promise<DiplomaResponse[]> {
-    const response = await fetch(`http://localhost:5258/api/diploma/many`, { // Ensure the URL matches
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(diplomasRequest)
-    });
-
-    if (!response.ok) {
-        if (response.status === 409) {
-            throw new Error("Some diplomas already exist!");
-        }
-        throw new Error("Failed to post diplomas!");
-    }
-
-    const result = await response.json() as DiplomaResponse[];
-    return result;
 }
